backend/Workspace101/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-import-export
python-dotenv
psycopg2-binary
google-genai
typesense
beautifulsoup4
ollama